feat: improve multiple choice dropdown a lot
