+++ components/Observation/MetadataWidget.vue
@@ -83,7 +83,7 @@
                 <USelectMenu
                   :name="field.label"
                   class="min-w-[200px]"
-                  :options="field.choices?.map((o) => ({ label: o }))"
+                  :options="getMultipleChoiceAddOptions(field as ProjectFieldResponse)"
                   v-model="state[field.label]"
                   :placeholder="field.required ? 'Select options or type freely' : 'Nothing picked'"
                   option-attribute="label"
@@ -94,7 +94,6 @@
                   creatable
                   by="label"
                   :disabled="!!$props.disabled"
-                  @update:model-value="(val: Array<{ label: string}>) => addCustomFieldToOptions(val, field)"
                 >
                   <template #option-create="{ option }">
                     <span class="flex-shrink-0 text-gray-400 text-xs">Custom:</span>
@@ -135,7 +134,7 @@
 
 <script lang="ts" setup>
   import type { FormError } from '@nuxt/ui/dist/runtime/types/form';
-  import { inputTypes, FieldType } from '~/utils/observationFields';
+  import { FieldType } from '~/utils/observationFields';
 
   const props = defineProps({
     observation: requireObservationProp,
@@ -143,38 +142,28 @@
     onSubmit: Function as PropType<Function>,
     disabled: Boolean as PropType<Boolean>,
     metadataDone: Boolean as PropType<Boolean>,
+    initialState: Object as PropType<any>,
+    inputs: requireProp<CMSInput[]>(),
   });
 
   const { params } = useRoute();
   const { sortFields } = await useProjects(params);
   const { patchObservation } = await useObservations(props.project.id);
 
+  function getMultipleChoiceAddOptions(field: ProjectFieldResponse) {
+    const updatedOptions = (field.choices || [])
+      .concat(getCustomFieldChoices(field, state))
+      .map((o) => ({ label: o }));
+    return updatedOptions;
+  }
+
   const form = ref();
-  const inputs = ref([] as CMSInput[]);
-  const state = ref(props.observation?.data as any);
+  const state = ref(Object.assign({ ...props.initialState }, props.observation?.data as any));
   const sortedFields = computed(() => sortFields(props.project));
-  const { report } = useSentry();
 
-  if (inputs.value.length == 0) {
-    buildForm(sortedFields.value);
-  }
 
-  function addCustomFieldToOptions(
-    value: Array<{ label: string}>,
-    field: { choices?: string[] }
-  ) {
-    if (!field.choices) {
-      // TODO: report error
-      return;
-    }
-
-    const newChoices = new Set(value.map(v => v.label).concat(field.choices));
-    if (newChoices.size === field.choices.length) {
-      return;
-    }
-    field.choices = Array.from(newChoices);
-  }
-
+  // TODO: validation function doesn't seem completely functional
+  //       - manuel edge-case testing required
   function validate(state: any): FormError[] {
     const errors = [] as FormError[];
 
@@ -244,89 +233,6 @@
     }
 
     return errors;
-  }
-
-  function buildForm(fields: ProjectFieldResponse[]) {
-    for (const field of fields) {
-
-      const useSimpleInput = Object.keys(inputTypes).includes(field.type);
-      const typ = field.type;
-
-      if (useSimpleInput) {
-        const inputArgs: CMSInputProps = {
-          placeholder: 'Enter ' + field.label,
-          name: field.label,
-          type: inputTypes[FieldType.STRING],
-        };
-
-        if (typ == FieldType.FLOAT) {
-          inputArgs.type = inputTypes[FieldType.FLOAT];
-          inputArgs.step = 0.1;
-        } else if (typ == FieldType.INT) {
-          inputArgs.type = inputTypes[FieldType.INT];
-        } else if (typ == FieldType.DATETIME) {
-          inputArgs.type = inputTypes[FieldType.DATETIME];
-        } else if (typ == FieldType.DATE) {
-          inputArgs.type = inputTypes[FieldType.DATE];
-        } else if (typ != FieldType.STRING) {
-          const err = new Error(`Field with type '${field.type}' is not support :( Try again in an hour`);
-          report('error', err);
-          throw err;
-        }
-
-        inputs.value.push({
-          field,
-          props: inputArgs,
-        });
-      } else {
-        if (typ == FieldType.BOOLEAN) {
-          inputs.value.push({
-            field,
-            props: {
-              label: field.label,
-              name: field.label,
-              type: 'checkbox',
-              checked: false,
-            } as CMSCheckboxProps,
-          });
-        } else if(typ == FieldType.TEXTAREA) {
-          inputs.value.push({
-            field,
-            props: {
-              name: field.label,
-            } as CMSTextAreaProps,
-          });
-        } else if (isMultipleChoice(typ)) {
-          if (!field.choices?.length) {
-            report('error', 'Radio button type has no values to pick from');
-            return;
-          }
-
-          // add arrays for multiple choices types
-          if (typ === 'MULTIPLE_CHOICE_ADD') {
-            if (!state.value[field.label]) {
-              state.value[field.label] = [];
-            } else {
-              // if custom choices are picked, add them to field.choices
-              const customChoices = state.value[field.label]
-                .map((v: { label: string }) => v.label)
-                .filter((v: string) => !field.choices.includes(v));
-              
-              field.choices = field.choices.concat(customChoices);
-            }
-          } 
-
-          inputs.value.push({
-            field,
-            props: {
-              name: field.label,
-            } as CMSMultipleChoiceProps,
-          });
-        } else {
-          report('warning', `Field with type '${field.type}' is not support :( Try again in an hour`);
-        }
-      }
-    }
   }
 
   async function submit() {
